# static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MaraBet AI API",
    "description": "API completa para análise preditiva de apostas esportivas com machine learning",
    "version": "1.0.0",
    "contact": {
      "name": "Equipe MaraBet AI",
      "email": "suporte@marabet.ai",
      "url": "https://marabet.ai"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Servidor de desenvolvimento"
    },
    {
      "url": "https://api.marabet.ai",
      "description": "Servidor de produção"
    }
  ],
...
  "paths": {
...
    "/api/predictions": {
      "get": {
        "tags": [
          "Predictions"
        ],
        "summary": "Listar Predições",
        "description": "Lista predições de partidas com paginação",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Número da página",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Itens por página",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "league_id",
            "in": "query",
            "description": "ID da liga para filtrar",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de predições",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionsListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Predictions"
        ],
        "summary": "Criar Predição",
        "description": "Cria uma nova predição para uma partida",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Predição criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/predictions/{match_id}": {
      "get": {
        "tags": [
          "Predictions"
        ],
        "summary": "Obter Predição",
        "description": "Obtém predição específica de uma partida",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "description": "ID da partida",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Predição encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Predição não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PredictionRequest": {
        "type": "object",
        "required": [
          "match_id",
          "home_team",
          "away_team",
          "league_id"
        ],
        "properties": {
          "match_id": {
            "type": "string",
            "example": "39_12345"
          },
          "home_team": {
            "type": "string",
            "example": "Manchester City"
          },
          "away_team": {
            "type": "string",
            "example": "Liverpool"
          },
          "league_id": {
            "type": "integer",
            "example": 39
          },
          "match_date": {
            "type": "string",
            "format": "date",
            "example": "2024-01-15"
          }
        }
      },
      "PredictionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "pred_123"
          },
          "match_id": {
            "type": "string",
            "example": "39_12345"
          },
          "home_team": {
            "type": "string",
            "example": "Manchester City"
          },
          "away_team": {
            "type": "string",
            "example": "Liverpool"
          },
          "predictions": {
            "type": "object",
            "properties": {
              "home_win": {
                "type": "number",
                "example": 0.45
              },
              "draw": {
                "type": "number",
                "example": 0.3
              },
              "away_win": {
                "type": "number",
                "example": 0.25
              }
            }
          },
          "confidence": {
            "type": "number",
            "example": 0.85
          },
          "expected_value": {
            "type": "number",
            "example": 0.12
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PredictionsListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PredictionResponse"
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer",
                "example": 1
              },
              "per_page": {
                "type": "integer",
                "example": 20
              },
              "total": {
                "type": "integer",
                "example": 100
              },
              "pages": {
                "type": "integer",
                "example": 5
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "validation_error"
              },
              "message": {
                "type": "string",
                "example": "Dados inválidos"
              },
              "details": {
                "type": "string",
                "example": "Campo obrigatório não fornecido"
              }
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Chave de API para autenticação"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ]
}